+++ ICD.Connect.Conferencing.Cisco/Devices/Codec/CiscoCodecDevice.cs
@@ -72,12 +72,6 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 		public event EventHandler<BoolEventArgs> OnConnectedStateChanged;
 
 		/// <summary>
-		/// Raised when the codec sends an error.
-		/// </summary>
-		[PublicAPI]
-		public event EventHandler<StringEventArgs> OnParsedError;
-
-		/// <summary>
 		/// System Configuration Commands
 		/// </summary>
 		private readonly string[] m_ConfigurationCommands =
@@ -131,6 +125,11 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 		public ePhonebookType PhonebookType { get; private set; }
 
 		/// <summary>
+		/// Determines which camera to use with PresenterTrack features.
+		/// </summary>
+		public int? PresenterTrackCameraId { get; set; }
+
+		/// <summary>
 		/// Provides the components attached to this codec.
 		/// </summary>
 		public CiscoComponentFactory Components { get { return m_Components; } }
@@ -169,7 +168,8 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 			Controls.Add(new CiscoCodecLayoutControl(this, 3));
 			Controls.Add(new CiscoCodecPresentationControl(this, 4));
 			Controls.Add(new CiscoCodecPowerControl(this, 5));
-			Controls.Add(new CiscoCalendarControl(this, 6));
+			Controls.Add(new CiscoCodecCalendarControl(this, 6));
+			Controls.Add(new CiscoCodecOccupancySensorControl(this, 7));
 		}
 
 		#endregion
@@ -201,7 +201,7 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
         [PublicAPI]
         public void SetPort(ISerialPort port)
         {
-            m_ConnectionStateManager.SetPort(port);
+	        m_ConnectionStateManager.SetPort(port);
         }
         
         /// <summary>
@@ -211,7 +211,6 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 		{
 			OnInitializedChanged = null;
 			OnConnectedStateChanged = null;
-			OnParsedError = null;
 
 			m_FeedbackTimer.Dispose();
 
@@ -264,6 +263,11 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 				command = string.Format(command, args);
 
 			m_ConnectionStateManager.Send(command + END_OF_LINE);
+
+#if !SIMPLSHARP
+			// Too fast!
+			ThreadingUtils.Sleep(10);
+#endif
 		}
 
 		/// <summary>
@@ -457,8 +461,6 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 			}
 
 			Log(eSeverity.Error, message);
-
-			OnParsedError.Raise(this, new StringEventArgs(message));
 		}
 
 		/// <summary>
@@ -670,6 +672,7 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 			settings.Port = m_ConnectionStateManager.PortNumber;
 			settings.PeripheralsId = PeripheralsId;
 			settings.PhonebookType = PhonebookType;
+			settings.PresenterTrackCameraId = PresenterTrackCameraId;
 		}
 
 		/// <summary>
@@ -681,8 +684,9 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 
 			PeripheralsId = null;
 			PhonebookType = ePhonebookType.Corporate;
+			PresenterTrackCameraId = null;
 
-			m_ConnectionStateManager.SetPort(null);
+			SetPort(null);
 		}
 
 		/// <summary>
@@ -696,6 +700,7 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 
 			PeripheralsId = settings.PeripheralsId;
 			PhonebookType = settings.PhonebookType;
+			PresenterTrackCameraId = settings.PresenterTrackCameraId;
 
 			ISerialPort port = null;
 
@@ -711,7 +716,7 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec
 				}
 			}
 
-			m_ConnectionStateManager.SetPort(port);
+			SetPort(port);
 		}
 
 		#endregion